+++ src/sagemaker/amazon/common.py
@@ -13,7 +13,6 @@
 """Placeholder docstring"""
 from __future__ import absolute_import
 
-import io
 import logging
 import struct
 import sys
@@ -21,74 +20,7 @@ import sys
 import numpy as np
 
 from sagemaker.amazon.record_pb2 import Record
-from sagemaker.deprecations import deprecated_class
-from sagemaker.deserializers import SimpleBaseDeserializer
-from sagemaker.serializers import SimpleBaseSerializer
 from sagemaker.utils import DeferredError
-
-
-class RecordSerializer(SimpleBaseSerializer):
-    """Serialize a NumPy array for an inference request."""
-
-    def __init__(self, content_type="application/x-recordio-protobuf"):
-        """Initialize a ``RecordSerializer`` instance.
-
-        Args:
-            content_type (str): The MIME type to signal to the inference endpoint when sending
-                request data (default: "application/x-recordio-protobuf").
-        """
-        super(RecordSerializer, self).__init__(content_type=content_type)
-
-    def serialize(self, data):
-        """Serialize a NumPy array into a buffer containing RecordIO records.
-
-        Args:
-            data (numpy.ndarray): The data to serialize.
-
-        Returns:
-            io.BytesIO: A buffer containing the data serialized as records.
-        """
-        if len(data.shape) == 1:
-            data = data.reshape(1, data.shape[0])
-
-        if len(data.shape) != 2:
-            raise ValueError(
-                "Expected a 1D or 2D array, but got a %dD array instead." % len(data.shape)
-            )
-
-        buffer = io.BytesIO()
-        write_numpy_to_dense_tensor(buffer, data)
-        buffer.seek(0)
-
-        return buffer
-
-
-class RecordDeserializer(SimpleBaseDeserializer):
-    """Deserialize RecordIO Protobuf data from an inference endpoint."""
-
-    def __init__(self, accept="application/x-recordio-protobuf"):
-        """Initialize a ``RecordDeserializer`` instance.
-
-        Args:
-            accept (union[str, tuple[str]]): The MIME type (or tuple of allowable MIME types) that
-                is expected from the inference endpoint (default:
-                "application/x-recordio-protobuf").
-        """
-        super(RecordDeserializer, self).__init__(accept=accept)
-
-    def deserialize(self, data, content_type):
-        """Deserialize RecordIO Protobuf data from an inference endpoint.
-
-        Args:
-            data (object): The protobuf message to deserialize.
-            content_type (str): The MIME type of the data.
-        Returns:
-            list: A list of records.
-        """
-        try:
-            return read_records(data)
-        finally:
-            data.close()
 
 
 def _write_feature_tensor(resolved_type, record, vector):
@@ -288,7 +220,3 @@ def _resolve_type(dtype):
     if dtype == np.dtype("float32"):
         return "Float32"
     raise ValueError("Unsupported dtype {} on array".format(dtype))
-
-
-numpy_to_record_serializer = deprecated_class(RecordSerializer, "numpy_to_record_serializer")
-record_deserializer = deprecated_class(RecordDeserializer, "record_deserializer")
